pyproject.toml
[tool.poetry]
name = "toolhub"
version = "0.1.0"
description = "An authoritative and well promoted catalog of Wikimedia tools"
authors = [
  "Bryan Davis <bd808@wikimedia.org>"
]
license = "GPL-3.0-or-later"
readme = "README.rst"
include = [
  "README.rst",
  "pyproject.toml",
  "COPYING"
]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^2.2"
django-environ = "^0.4.5"
django-log-request-id = "^1.6.0"
mysqlclient = "^2.0.1"
social-auth-app-django = "^4.0.0"
django-webpack-loader = "^0.7.0"
djangorestframework = "^3.12.1"
PyYAML = "^5.3.1"
uritemplate = "^3.0.1"
django-filter = "^2.4.0"
jsonfield = "^3.1.0"
drf-friendly-errors = {git = "https://github.com/bd808/drf-friendly-errors.git", rev = "933eb5b"}
drf-spectacular = "^0.13.1"
requests = "^2.25.0"
drf-nested-routers = "^0.92.1"
whitenoise = {extras = ["brotli"], version = "^5.2.0"}
django-oauth-toolkit = "^1.3.3"
jsonschema = "^3.2.0"
django-reversion = "^3.0.9"
spdx-license-list = "^0.5.2"
jsonpatch = "^1.28"
django-reversion-compare = "^0.13.1"
elasticsearch-dsl = ">=6.0.0,<7.0.0"
django-elasticsearch-dsl = ">=6.0.0,<7.0.0"
django-elasticsearch-dsl-drf = "^0.21"
django-safedelete = "^1.0.0"
rules = "^2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flakehell = "^0.6.0"
black = "^20.8b1"
flake8-blind-except = "^0.1.1"
flake8-bugbear = "^20.1.4"
flake8-builtins = "^1.5.3"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
flake8-logging-format = "^0.6.0"
flake8-rst-docstrings = "^0.0.13"
flake8-comprehensions = "^3.2.3"
pytest-cov = "^2.10.1"
django-coverage-plugin = "^1.8.0"
bandit = "^1.6.2"
sphinx = "^3.2.1"
sphinxcontrib-django = "^0.5.1"
sphinx_rtd_theme = "^0.5.0"
coverage = {extras = ["toml"], version = "^5.3"}
httpretty = "^1.0.2"
unittest2 = "^1.1.0"
pylint-django = "^2.3.0"
requests-mock = "^1.8.0"

[tool.flakehell]
exclude = [
    ".git",
    "__pycache__",
    "docs",
    "manage.py",
    "migrations",
    "node_modules",
]
format = "grouped"
max_line_length = 79
show_source = true

[tool.flakehell.plugins]
"flake8-*" = ["+*"]
flake8-docstrings = ["+*", "-D100", "-D105", "-D400", "-D401"]
pycodestyle = ["+*", "-E501", "-W503", "-W504"]
pyflakes = ["+*"]
pylint = [
    "+*",
    "-C*",
    "-E1101",
    "-R0201",
    "-R0901",
    "-R0903",
    "-W0235",
    "-W0511",
]

[tool.pylint]

[tool.pylint.master]
load-plugins = "pylint_django"

[tool.pylint.DESIGN]
max-branches = 16

[tool.black]
line-length = 79
target_version = ['py37']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
known_first_party = ["toolhub", "vue"]
multi_line_output = 3
lines_after_imports = 2
force_grid_wrap = 0
combine_as_imports = true
include_trailing_comma = true

[tool.coverage.run]
source = ["toolhub", "vue"]
omit = [
  "*/migrations/*",
  "toolhub/wsgi.py",
  "toolhub/apps/toolinfo/schema2model.py",
]
branch = true
dynamic_context = "test_function"
relative_files = true

[tool.coverage.report]
fail_under = 85
show_missing = true
skip_covered = true
skip_empty = true
ignore_errors = true
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
]

[tool.coverage.html]
directory = "docs/htmlcov"
show_contexts = true

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = true
envlist = py37

[testenv]
whitelist_externals = poetry
setenv=
    PYTHONDONTWRITEBYTECODE=true
    PYTHONPATH={toxinidir}
    DJANGO_SECRET_KEY="this is not really a secret"
commands=
  poetry check
  poetry install -v --no-root
  poetry run flakehell lint
  poetry run black --check --diff .
  poetry run coverage erase
  poetry run coverage run --branch manage.py test
  poetry run coverage report
  poetry run coverage html
  poetry run bandit -ii -r toolhub/
"""

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
